Move multipart writer code into rest package
