+++ db/attachment.go
@@ -10,29 +10,12 @@
 package db
 
 import (
-	"bytes"
-	"compress/gzip"
-	"crypto/md5"
 	"crypto/rand"
 	"crypto/sha1"
 	"encoding/base64"
-	"fmt"
-	"io"
-	"io/ioutil"
-	"mime/multipart"
-	"net/http"
-	"net/textproto"
-	"strings"
 
 	"github.com/couchbase/sync_gateway/base"
 )
-
-// AttachmentsMeta shorter than this will be left in the JSON as base64 rather than being a separate
-// MIME part.
-const kMaxInlineAttachmentSize = 200
-
-// JSON bodies smaller than this won't be GZip-encoded.
-const kMinCompressedJSONSize = 300
 
 // Key for retrieving an attachment from Couchbase.
 type AttachmentKey string
@@ -67,7 +50,7 @@ func (db *Database) storeAttachments(doc *Document, newAttachmentsMeta Attachmen
 		data := meta["data"]
 		if data != nil {
 			// Attachment contains data, so store it in the db:
-			attachment, err := decodeAttachment(data)
+			attachment, err := DecodeAttachment(data)
 			if err != nil {
 				return nil, err
 			}
@@ -211,242 +194,6 @@ func (db *Database) setAttachments(attachments AttachmentData) error {
 	return nil
 }
 
-//////// MIME MULTIPART:
-
-// Parses a JSON MIME body, unmarshaling it into "into".
-func ReadJSONFromMIME(headers http.Header, input io.Reader, into interface{}) error {
-	contentType := headers.Get("Content-Type")
-	if contentType != "" && !strings.HasPrefix(contentType, "application/json") {
-		return base.HTTPErrorf(http.StatusUnsupportedMediaType, "Invalid content type %s", contentType)
-	}
-
-	switch headers.Get("Content-Encoding") {
-	case "gzip":
-		var err error
-		if input, err = gzip.NewReader(input); err != nil {
-			return err
-		}
-	case "":
-		break
-	default:
-		return base.HTTPErrorf(http.StatusUnsupportedMediaType, "Unsupported Content-Encoding; use gzip")
-	}
-
-	decoder := base.JSONDecoder(input)
-	decoder.UseNumber()
-	if err := decoder.Decode(into); err != nil {
-		base.Warnf(base.KeyAll, "Couldn't parse JSON in HTTP request: %v", err)
-		return base.HTTPErrorf(http.StatusBadRequest, "Bad JSON")
-	}
-	return nil
-}
-
-type attInfo struct {
-	name        string
-	contentType string
-	data        []byte
-}
-
-func writeJSONPart(writer *multipart.Writer, contentType string, body Body, compressed bool) (err error) {
-	bytes, err := base.JSONMarshal(body)
-	if err != nil {
-		return err
-	}
-	if len(bytes) < kMinCompressedJSONSize {
-		compressed = false
-	}
-
-	partHeaders := textproto.MIMEHeader{}
-	partHeaders.Set("Content-Type", contentType)
-	if compressed {
-		partHeaders.Set("Content-Encoding", "gzip")
-	}
-	part, err := writer.CreatePart(partHeaders)
-	if err != nil {
-		return err
-	}
-
-	if compressed {
-		gz := gzip.NewWriter(part)
-		_, err = gz.Write(bytes)
-		gz.Close()
-	} else {
-		_, err = part.Write(bytes)
-	}
-	return
-}
-
-// Writes a revision to a MIME multipart writer, encoding large attachments as separate parts.
-func (db *Database) WriteMultipartDocument(body Body, writer *multipart.Writer, compress bool) {
-	// First extract the attachments that should follow:
-	following := []attInfo{}
-	for name, value := range GetBodyAttachments(body) {
-		meta := value.(map[string]interface{})
-		if meta["stub"] != true {
-			var err error
-			var info attInfo
-			info.contentType, _ = meta["content_type"].(string)
-			info.data, err = decodeAttachment(meta["data"])
-			if info.data == nil {
-				base.WarnfCtx(db.Ctx, base.KeyAll, "Couldn't decode attachment %q of doc %q: %v", base.UD(name), base.UD(body[BodyId]), err)
-				meta["stub"] = true
-				delete(meta, "data")
-			} else if len(info.data) > kMaxInlineAttachmentSize {
-				info.name = name
-				following = append(following, info)
-				meta["follows"] = true
-				delete(meta, "data")
-			}
-		}
-	}
-
-	// Write the main JSON body:
-	writeJSONPart(writer, "application/json", body, compress)
-
-	// Write the following attachments
-	for _, info := range following {
-		partHeaders := textproto.MIMEHeader{}
-		if info.contentType != "" {
-			partHeaders.Set("Content-Type", info.contentType)
-		}
-		partHeaders.Set("Content-Disposition", fmt.Sprintf("attachment; filename=%q", info.name))
-		part, _ := writer.CreatePart(partHeaders)
-		db.DatabaseContext.DbStats.StatsCblReplicationPull().Add(base.StatKeyAttachmentPullCount, 1)
-		db.DatabaseContext.DbStats.StatsCblReplicationPull().Add(base.StatKeyAttachmentPullBytes, int64(len(info.data)))
-		part.Write(info.data)
-
-	}
-}
-
-// Adds a new part to the given multipart writer, containing the given revision.
-// The revision will be written as a nested multipart body if it has attachments.
-func (db *Database) WriteRevisionAsPart(revBody Body, isError bool, compressPart bool, writer *multipart.Writer) error {
-	partHeaders := textproto.MIMEHeader{}
-	docID, _ := revBody[BodyId].(string)
-	revID, _ := revBody[BodyRev].(string)
-	if len(docID) > 0 {
-		partHeaders.Set("X-Doc-ID", docID)
-		partHeaders.Set("X-Rev-ID", revID)
-	}
-
-	if hasInlineAttachments(revBody) {
-		// Write as multipart, including attachments:
-		// OPT: Find a way to do this w/o having to buffer the MIME body in memory!
-		var buffer bytes.Buffer
-		docWriter := multipart.NewWriter(&buffer)
-		contentType := fmt.Sprintf("multipart/related; boundary=%q",
-			docWriter.Boundary())
-		partHeaders.Set("Content-Type", contentType)
-		db.WriteMultipartDocument(revBody, docWriter, compressPart)
-		docWriter.Close()
-		content := bytes.TrimRight(buffer.Bytes(), "\r\n")
-
-		part, err := writer.CreatePart(partHeaders)
-		if err == nil {
-			_, err = part.Write(content)
-		}
-		return err
-	} else {
-		// Write as JSON:
-		contentType := "application/json"
-		if isError {
-			contentType += `; error="true"`
-		}
-		return writeJSONPart(writer, contentType, revBody, compressPart)
-	}
-}
-
-func ReadMultipartDocument(reader *multipart.Reader) (Body, error) {
-	// First read the main JSON document body:
-	mainPart, err := reader.NextPart()
-	if err != nil {
-		return nil, err
-	}
-	var body Body
-	err = ReadJSONFromMIME(http.Header(mainPart.Header), mainPart, &body)
-	mainPart.Close()
-	if err != nil {
-		return nil, err
-	}
-
-	// Collect the attachments with a "follows" property, which will appear as MIME parts:
-	followingAttachments := map[string]map[string]interface{}{}
-	for name, value := range GetBodyAttachments(body) {
-		if meta := value.(map[string]interface{}); meta["follows"] == true {
-			followingAttachments[name] = meta
-		}
-	}
-
-	// Subroutine to look up a following attachment given its digest. (I used to precompute a
-	// map from digest->name, which was faster, but that broke down if there were multiple
-	// attachments with the same contents! See #96)
-	findFollowingAttachment := func(withDigest string) (string, map[string]interface{}) {
-		for name, meta := range followingAttachments {
-			if meta["follows"] == true {
-				if digest, ok := meta["digest"].(string); ok && digest == withDigest {
-					return name, meta
-				}
-			}
-		}
-		return "", nil
-	}
-
-	// Read the parts one by one:
-	for i := 0; i < len(followingAttachments); i++ {
-		part, err := reader.NextPart()
-		if err != nil {
-			if err == io.EOF {
-				err = base.HTTPErrorf(http.StatusBadRequest,
-					"Too few MIME parts: expected %d attachments, got %d",
-					len(followingAttachments), i)
-			}
-			return nil, err
-		}
-		data, err := ioutil.ReadAll(part)
-		part.Close()
-		if err != nil {
-			return nil, err
-		}
-
-		// Look up the attachment by its digest:
-		digest := Sha1DigestKey(data)
-		name, meta := findFollowingAttachment(digest)
-		if meta == nil {
-			name, meta = findFollowingAttachment(Md5DigestKey(data))
-			if meta == nil {
-				return nil, base.HTTPErrorf(http.StatusBadRequest,
-					"MIME part #%d doesn't match any attachment", i+2)
-			}
-		}
-
-		length, ok := base.ToInt64(meta["encoded_length"])
-		if !ok {
-			length, ok = base.ToInt64(meta["length"])
-		}
-		if ok {
-			if length != int64(len(data)) {
-				return nil, base.HTTPErrorf(http.StatusBadRequest, "Attachment length mismatch for %q: read %d bytes, should be %d", name, len(data), length)
-			}
-		}
-
-		// Stuff the data into the attachment metadata and remove the "follows" property:
-		delete(meta, "follows")
-		meta["data"] = data
-		meta["digest"] = digest
-	}
-
-	// Make sure there are no unused MIME parts:
-	if _, err = reader.NextPart(); err != io.EOF {
-		if err == nil {
-			err = base.HTTPErrorf(http.StatusBadRequest,
-				"Too many MIME parts (expected %d)", len(followingAttachments)+1)
-		}
-		return nil, err
-	}
-
-	return body, nil
-}
-
 type AttachmentCallback func(name string, digest string, knownData []byte, meta map[string]interface{}) ([]byte, error)
 
 // Given a document body, invokes the callback once for each attachment that doesn't include
@@ -530,20 +277,11 @@ func AttachmentDigests(attachments AttachmentsMeta) []string {
 	return digests
 }
 
-func hasInlineAttachments(body Body) bool {
-	for _, value := range GetBodyAttachments(body) {
-		if meta, ok := value.(map[string]interface{}); ok && meta["data"] != nil {
-			return true
-		}
-	}
-	return false
-}
-
 func attachmentKeyToString(key AttachmentKey) string {
 	return base.AttPrefix + string(key)
 }
 
-func decodeAttachment(att interface{}) ([]byte, error) {
+func DecodeAttachment(att interface{}) ([]byte, error) {
 	switch att := att.(type) {
 	case []byte:
 		return att, nil
@@ -558,11 +296,4 @@ func Sha1DigestKey(data []byte) string {
 	digester := sha1.New()
 	digester.Write(data)
 	return "sha1-" + base64.StdEncoding.EncodeToString(digester.Sum(nil))
-}
-
-// This is only here for backwards compatibility.  Otherwise should be avoided.
-func Md5DigestKey(data []byte) string {
-	digester := md5.New()
-	digester.Write(data)
-	return "md5-" + base64.StdEncoding.EncodeToString(digester.Sum(nil))
 }
